+++ src/extension.ts
@@ -348,7 +348,7 @@ async function onDidChangeTextEditorSelection(event: vscode.TextEditorSelectionC
     repo.files.clear();
     loadFile(repo, editor, { reuse: file });
   }
-  const line = event.selections[0].start.line;
+  const line = event.selections[0].active.line;
   // Use MAX_VALUE so that it will go on the end of the line even if you type
   // more characters before the decoration gets applied.
   const end = new vscode.Position(line, Number.MAX_VALUE);